Upload testnet block

+++ src/chainparams.h
@@ -97,7 +97,7 @@ public:
     CBaseChainParams::Network NetworkID() const { return networkID; }
     int NewMasternodeReward_StartBlock() const { return nNewMasternodeReward_StartBlock; }
     int NewMasternodeReward_Collateral() const { return nNewMasternodeReward_Collateral; }
-    float NewMasternodeReward_MNPercent() const { return nNewMasternodeReward_MNPercent; }
+    double NewMasternodeReward_MNPercent() const { return nNewMasternodeReward_MNPercent; }
 
 protected:
     CChainParams() {}
@@ -143,7 +143,7 @@ protected:
     int64_t nBudget_Fee_Confirmations;
     int nNewMasternodeReward_StartBlock;
     int nNewMasternodeReward_Collateral;
-    float nNewMasternodeReward_MNPercent;
+    double nNewMasternodeReward_MNPercent;
 };
 
 /**
